With-Credentials
Added ability to pass PSCredentials to the script so that this can be ran using alternative credentials against a local or remote machine.  Also added the ability to pass a UserName which will then prompt for a PassWord.

Also modified the script to work against remote older Windows systems such as Windows 2000/XP/2003 that are NOT in the same domain.